update :: ChangeEventStatus로 변경
