+++ src/main/java/com/mindway/server/v2/domain/event/scheduler/EventScheduler.java
@@ -1,7 +1,6 @@
 package com.mindway.server.v2.domain.event.scheduler;
 
-import com.mindway.server.v2.domain.event.service.ChangeNowToPastService;
-import com.mindway.server.v2.domain.event.service.ChangePendingToNowService;
+import com.mindway.server.v2.domain.event.service.ChangeEventStatusService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -10,17 +9,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class EventScheduler {
 
-    private final ChangeNowToPastService changeNowToPastService;
-    private final ChangePendingToNowService changePendingToNowService;
+    private final ChangeEventStatusService changeEventStatusService;
 
     @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
-    public void changeNowToPast() {
-        changeNowToPastService.execute();
+    public void changeEventStatus() {
+        changeEventStatusService.execute();
     }
-
-    @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
-    public void changePendingToNow() {
-        changePendingToNowService.execute();
-    }
-
 }
